Train a new FastRet model (FRM) for retention time prediction

Trains a new model from molecule SMILES to predict retention times (RT) using the specified method.

Usage

train_frm(
  df = read_rp_xlsx(),
  method = "lasso",
  verbose = 1,
  nfolds = 5,
  nw = 1,
  degree_polynomial = 1,
  interaction_terms = FALSE,
  rm_near_zero_var = TRUE,
  rm_na = TRUE,
  rm_ns = FALSE,
  seed = NULL
)

Arguments

df

A dataframe with columns "NAME", "RT", "SMILES" and optionally a set of chemical descriptors. If no chemical descriptors are provided, they are calculated using the function preprocess_data().

method

A string representing the prediction algorithm. Either "lasso", "ridge" or "gbtree".

verbose

A logical value indicating whether to print progress messages.

nfolds

An integer representing the number of folds for cross validation.

nw

An integer representing the number of workers for parallel processing.

degree_polynomial

An integer representing the degree of the polynomial. Polynomials up to the specified degree are included in the model.

interaction_terms

A logical value indicating whether to include interaction terms in the model.

rm_near_zero_var

A logical value indicating whether to remove near zero variance predictors. Setting this to TRUE can cause the CV results to be overoptimistic, as the variance filtering is done on the whole dataset, i.e. information from the test folds is used for feature selection.

rm_na

A logical value indicating whether to remove NA values. Setting this to TRUE can cause the CV results to be overoptimistic, as the variance filtering is done on the whole dataset, i.e. information from the test folds is used for feature selection.

rm_ns

A logical value indicating whether to remove chemical descriptors that were considered as not suitable for linear regression based on an analysis of the Retip::HILIC dataset. See check_lm_suitability() for details on the analysis.

seed

An integer value to set the seed for random number generation to allow for reproducible results.

Value

A trained FastRet model.

Details

Setting rm_near_zero_var and/or rm_na to TRUE can cause the CV results to be overoptimistic, as the predictor filtering is done on the whole dataset, i.e. information from the test folds is used for feature selection.

Examples

system.time(m <- train_frm(RP[1:80, ], method = "lasso", nfolds = 2, nw = 1, verbose = 0))
#> 2024-06-13 11:39:22.37 Calculating chemical descriptors for: 'C1=NC2=C(C(=N1)N)N=CN2C3C(C(C(O3)COP(=O)(O)OP(=O)(O)OCC4C(C(C(O4)O)O)O)O)O'
#> 2024-06-13 11:39:22.69 Calculating chemical descriptors for: 'CC(=O)NCCCNCCCCNCCCN'
#> 2024-06-13 11:39:22.80 Calculating chemical descriptors for: '[C@@H]1(C[C@H](O)[C@@H](CO)O1)N2C(=O)NC(=O)C=C2'
#> 2024-06-13 11:39:22.92 Calculating chemical descriptors for: 'CC1=CN(C(=O)NC1=O)C2CC(C(O2)COP(=O)(O)O)O'
#> 2024-06-13 11:39:23.08 Calculating chemical descriptors for: '[P](O)(=O)(O[P](=O)(O)O)OC[C@@H]1[C@H](C[C@@H](O1)N2C(=O)N=C(N)C=C2)O'
#> 2024-06-13 11:39:23.26 Calculating chemical descriptors for: 'CC1=CC2=C(C=C1C)N(C3=NC(=O)NC(=O)C3=N2)CC(C(C(COP(=O)(O)OP(=O)(O)OCC4C(C(C(O4)N5C=NC6=C5N=CN=C6N)O)O)O)O)O'
#> 2024-06-13 11:39:23.92 Calculating chemical descriptors for: 'C(C1C(C(C(C(O1)O)N)O)O)OP(=O)(O)O'
#> 2024-06-13 11:39:24.05 Calculating chemical descriptors for: 'C(CCNCCCN)CNCCCN'
#> 2024-06-13 11:39:24.15 Calculating chemical descriptors for: 'C(CCNCCCN)CN'
#> 2024-06-13 11:39:24.25 Calculating chemical descriptors for: 'C(C(C1C(=O)C(=O)C(=O)O1)O)O'
#> 2024-06-13 11:39:24.36 Calculating chemical descriptors for: '[C@@H]14O[P](OC[C@H]1O[C@@H]([N]2C3=C(N=C2)C(=O)N=C(N3)N)[C@@H]4O)(O)=O'
#> 2024-06-13 11:39:24.60 Calculating chemical descriptors for: 'C(C(C(C(C(C(=O)O)O)O)O)O)OP(=O)(O)O'
#> 2024-06-13 11:39:24.72 Calculating chemical descriptors for: 'C[N+](C)(C)CCCCC(C(=O)[O-])N'
#> 2024-06-13 11:39:24.82 Calculating chemical descriptors for: 'C1=NC2=C(C(=N1)N)N=CN2C3C(C(C(O3)COP(=O)(O)OP(=O)(O)O)O)O'
#> 2024-06-13 11:39:25.06 Calculating chemical descriptors for: 'C(C(=O)C(=O)O)O'
#> 2024-06-13 11:39:25.16 Calculating chemical descriptors for: 'C1=NC2=C(N1C3C(C(C(O3)COP(=O)(O)O)O)O)NC(=NC2=O)N'
#> 2024-06-13 11:39:25.37 Calculating chemical descriptors for: '[C@@H]1(C[C@H](O)[C@@H](CO)O1)N2C(=O)N=C(N)C=C2'
#> 2024-06-13 11:39:25.50 Calculating chemical descriptors for: 'C(C1C(C(C(C(O1)O)O)O)O)OP(=O)(O)O'
#> 2024-06-13 11:39:25.62 Calculating chemical descriptors for: 'C1C(C(OC1N2C=CC(=O)NC2=O)COP(=O)(O)O)O'
#> 2024-06-13 11:39:25.77 Calculating chemical descriptors for: 'C[N+](C)(CCOP([O-])(OP(O)(OCC1C(O)C(O)C(N2C=CC(N)=NC2=O)O1)=O)=O)C'
#> 2024-06-13 11:39:26.00 Calculating chemical descriptors for: 'C(CCN)CN'
#> 2024-06-13 11:39:26.11 Calculating chemical descriptors for: 'CC1=NC=C(C(=C1O)CN)CO'
#> 2024-06-13 11:39:26.22 Calculating chemical descriptors for: 'C(CCN=C(N)N)CN'
#> 2024-06-13 11:39:26.31 Calculating chemical descriptors for: 'C1C(C(OC1N2C=NC3=C2NC(=NC3=O)N)COP(=O)(O)OP(=O)(O)OP(=O)(O)O)O'
#> 2024-06-13 11:39:26.57 Calculating chemical descriptors for: 'CC1=NC=C(C(=C1O)C=O)COP(=O)(O)O'
#> 2024-06-13 11:39:26.69 Calculating chemical descriptors for: '[C@H]34[C@H]2[C@@H]([C@@]1(C(=CC(=O)CC1)CC2)C)CC[C@@]3([C@](C(CO)=O)(O)CC4)C'
#> 2024-06-13 11:39:27.09 Calculating chemical descriptors for: 'C1=NC3=C([N]1C2OC(CO[P](O[P](=O)(O)O)(=O)O)C(C2)O)N=CN=C3N'
#> 2024-06-13 11:39:27.31 Calculating chemical descriptors for: 'CN(CC(=O)O)C(=NP(=O)(O)O)N'
#> 2024-06-13 11:39:27.41 Calculating chemical descriptors for: 'C(C1C(C(C(C(O1)O)N)O)O)O'
#> 2024-06-13 11:39:27.53 Calculating chemical descriptors for: 'CC1=NC=C(C(=C1O)C=O)CO'
#>    user  system elapsed 
#>   8.399   0.080   5.747 
# For the sake of a short runtime, only the first 80 rows of the RP dataset
# are used in this example. In practice, you should always use the entire
# training dataset for model training.