Get Chemical Descriptors for a list of molecules

Calculate Chemical Descriptors for a list of molecules. Molecules can appear multiple times in the list.

Usage

getCDs(df, verbose = 1, nw = 1)

Arguments

df

dataframe with two mandatory columns: "NAME" and "SMILES"

verbose

0: no output, 1: progress, 2: more progress and warnings

nw

number of workers for parallel processing

Value

A dataframe with the chemical descriptor values appended as columns to the input dataframe.

Examples

cds <- getCDs(head(RP, 3), verbose = 1, nw = 1)
#> 2025-02-07 19:06:10.34 Obtaining chemical descriptors for a dataframe of dimension 3 x 3
#> 2025-02-07 19:06:10.34 Calculating chemical descriptors
#> 2025-02-07 19:06:10.34 Initializing RAM cache for chemical descriptors
#> 2025-02-07 19:06:10.35 Loading chemical descriptors for 'CCC(C(=O)O)O' from RAM cache
#> 2025-02-07 19:06:10.35 Loading chemical descriptors for 'COC1=C(C=CC(=C1)CCN)O' from RAM cache
#> 2025-02-07 19:06:10.36 Loading chemical descriptors for 'C1=NC2=C(N1)C(=NC=N2)N' from RAM cache
#> 2025-02-07 19:06:10.36 Collecting results
#> 2025-02-07 19:06:10.36 Finished calculating chemical descriptors in 0.01955104 secs